Added public GetError method to the packet.
Before the only way for the user to get the error was to call
GetResultObject. However, the parse would also be called in this
case and then the user wouldn't know (programatically) where the error
came from. Off course, the librarie's user could make some string
matching to try to deduce, but this is error prone and non scalable.